Python/requirements.txt
absl-py==1.0.0
astor==0.8.1
backcall==0.2.0
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
fonttools==4.33.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.46.3
h5py==2.10.0
importlib-metadata==4.11.4
jedi==0.18.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.18.5
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.19.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.7.3
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
